+++ admin/esercizi.js
@@ -255,15 +255,6 @@
     });
   }
 
-  function escapeHtml(value) {
-    return String(value || '')
-      .replace(/&/g, '&amp;')
-      .replace(/</g, '&lt;')
-      .replace(/>/g, '&gt;')
-      .replace(/"/g, '&quot;')
-      .replace(/'/g, '&#39;');
-  }
-
   function slugify(value) {
     return String(value || '')
       .toLowerCase()